requirements.txt
pdfplumber==0.10.2
scikit-learn==1.5.0
nltk==3.8.1
python-dotenv==1.0.1
tqdm==4.66.4
joblib==1.4.2
google-generativeai==0.8.0
greenlet==3.1.1
playwright==1.54.0
python-docx>=1.1.0
